feat: allow grafana loki search configuration
